+++ src/scripts/distancemat.py
@@ -39,7 +39,7 @@ def debug(msg, *args, logger=None):
 
 @logwrap
 def warn(msg, *args, logger=None):
-    logger.debug(msg, *args)
+    logger.warn(msg, *args)
 
 
 class DistanceMatrix():
@@ -47,6 +47,13 @@ class DistanceMatrix():
     def __init__(self):
         self._entry_fmt = '<2h1f'
         self.dist_dict = {}
+
+    def __iter__(self):
+        def _iterator():
+            for k0, ref_dict in self.dist_dict.items():
+                for k1, dist in ref_dict.items():
+                    yield (k0, k1), dist
+        return _iterator()
 
     def __setitem__(self, eids, val):
         k0, k1 = eids
@@ -155,13 +162,23 @@ eman = read_into_manager(jsonf, strip=True)
 info('Loading cluster information')
 dframe = pd.read_csv(clusf, skiprows=range(1, 3))
 
+cluster_ids = set([])
+clst_key = ('clst', 1)
+invalid = set([])
 for ent in eman:
     try:
         cluster, = dframe[dframe.CellID == ent.eid].Cluster
-        ent.tags.add(cluster)
+        # ent.tags.add(cluster)
+        ent.scalars[clst_key] = int(cluster[1:])
+        cluster_ids.add(int(cluster[1:]))
     except ValueError:
         warn('no cluster assignment for %d', ent.eid)
+        ent.isActive = False
+        invalid.add(ent.eid)
         continue
+for inv_eid in invalid:
+    assert eman.popEntity(inv_eid) == inv_eid
+info('Using %d cluster: %s', len(cluster_ids), str(cluster_ids))
 
 info('Loading distance matrix')
 distmat = DistanceMatrix()
@@ -169,13 +186,30 @@ distmat.load(distf)
 
 info('Building neighborhood graph')
 graph = nx.Graph()
-graph.add_nodes_from()
-graph.add_edges_from([(1, 2), (1, 3), (2, 4), (2, 3)])
+for ent in eman.iter_active():
+    poly = Polygon(ent.contours[0])
+    ppos = tuple(*poly.centroid.coords)
+    props = dict(
+        cluster=ent.scalars[clst_key],
+        pos=ppos,)
+    graph.add_node(ent.eid, **props)
 
+for edge, dist in distmat:
+    if dist <= 1:
+        graph.add_edge(*edge, distance=dist)
+
+# n0 = len(graph.nodes)
+# isolated = [n for n, d in iter(graph.degree) if not d]
+# graph.remove_nodes_from(isolated)
+# n1 = len(graph.nodes)
+# info('Removed %d nodes', n0 - n1)
+
+info('Plotting')
 f, ax = plt.subplots()
-nx.draw(graph, with_labels=True, ax=ax)
+npos = {}
+for node in graph.nodes:
+    npos[node] = graph.nodes[node].get('pos')
+nx.draw_networkx(graph, pos=npos)#, with_labels=True, ax=ax)
 plt.show()
-
-
 
 ip.embed()